Plot 3D free energy surface object

`plot.fes3d` plots 3D free energy surface using .

# S3 method for fes3d
plot(x, xlab = NULL, ylab = NULL, zlab = NULL, 
  xlim = NULL, ylim = NULL, zlim = NULL, level = NULL,
  col = NULL, alpha = NULL, main = NULL, sub = NULL, fill = TRUE, ...)

Arguments

x

fes3d object.

xlab

a title for the x axis: see 'title'.

ylab

a title for the y axis: see 'title'.

zlab

a title for the z axis: see 'title'.

xlim

numeric vector of length 2, giving the x coordinates range.

ylim

numeric vector of length 2, giving the y coordinates range.

zlim

numeric vector of length 2, giving the z coordinates range.

main

an overall title for the plot: see 'title'.

sub

a sub title for the plot: see 'title'.

level

number or numeric vector of levels at which to draw 3D isosurface.

col

color of the free energy surface. It can be a single color or a vector with multiple colors for multiple 3D isosurfaces.

alpha

number or numeric vector of alpha levels (transparency) of 3D isosurfaces.

fill

a logical value indicating whether 3D isosurface is ploted as solid surface (True) or wireframe (False).

...

further arguments passed to or from other methods.

Examples

tfes3d<-fes(acealanme3d, imax=5000)
plot(tfes3d)