Plot minima3d object

`plot.minima3d` plots 3D free energy surface with minima. The free energy surface is plotted the same way as by plot.fes3d with additional minima labels.

# S3 method for minima3d
plot(x, xlab = "CV1", ylab = "CV2", zlab = "CV3", level = NULL,
  col = NULL, alpha = NULL, main = NULL, sub = NULL, fill = TRUE, ...)

Arguments

x

minima3d object.

main

an overall title for the plot: see 'title'.

sub

a sub title for the plot: see 'title'.

xlab

a title for the x axis: see 'title'.

ylab

a title for the y axis: see 'title'.

zlab

a title for the z axis: see 'title'.

level

number or numeric vector of levels at which to draw 3D isosurface.

col

color of the free energy surface. It can be a single color or a vector with multiple colors for multiple 3D isosurfaces.

alpha

number or numeric vector of alpha levels (transparency) of 3D isosurfaces.

fill

a logical value indicating whether 3D isosurface is ploted as solid surface (True) or wireframe (False).

...

further arguments passed to or from other methods.

Examples

tfes<-fes(acealanme3d, imax=5000)
minima<-fesminima(tfes)
plot(minima)